chore(ml): harden worklets and tests
