+++ app/src/services/hybridDefaults.ts
@@ -1,0 +1,7 @@
+export const HYBRID_DEFAULTS = {
+  localThreshold: 0.6,
+  cloudThreshold: 0.8,
+  remoteTimeoutMs: 400,
+  maxConsecutiveTimeouts: 3,
+  cooldownMs: 10 * 60 * 1000,
+};
